+++ cmd/geoipupdate/config.go
@@ -26,7 +26,7 @@ type Config struct {
 }
 
 // NewConfig parses the configuration file.
-func NewConfig(
+func NewConfig( // nolint: gocyclo
 	file,
 	databaseDirectory string,
 ) (*Config, error) {
@@ -74,7 +74,7 @@ func NewConfig(
 			keysSeen["AccountID"] = struct{}{}
 			keysSeen["UserId"] = struct{}{}
 		case "DatabaseDirectory":
-			config.DatabaseDirectory = value
+			config.DatabaseDirectory = filepath.Clean(value)
 		case "EditionIDs", "ProductIds":
 			config.EditionIDs = strings.Fields(value)
 			keysSeen["EditionIDs"] = struct{}{}
@@ -84,7 +84,7 @@ func NewConfig(
 		case "LicenseKey":
 			config.LicenseKey = value
 		case "LockFile":
-			config.LockFile = value
+			config.LockFile = filepath.Clean(value)
 		case "PreserveFileTimes":
 			if value != "0" && value != "1" {
 				return nil, errors.New("`PreserveFileTimes' must be 0 or 1")
@@ -132,11 +132,11 @@ func NewConfig(
 
 	// Argument takes precedence.
 	if databaseDirectory != "" {
-		config.DatabaseDirectory = databaseDirectory
+		config.DatabaseDirectory = filepath.Clean(databaseDirectory)
 	}
 
 	if config.DatabaseDirectory == "" {
-		config.DatabaseDirectory = defaultDatabaseDirectory
+		config.DatabaseDirectory = filepath.Clean(defaultDatabaseDirectory)
 	}
 
 	if host == "" {
@@ -195,7 +195,7 @@ func parseProxy(
 		return nil, errors.Wrap(err, "error parsing proxy URL")
 	}
 
-	if strings.Index(u.Host, ":") == -1 {
+	if !strings.Contains(u.Host, ":") {
 		u.Host += ":1080" // The 1080 default historically came from cURL.
 	}